+++ packages/okta-auth-js/test/karma/spec/loginFlow.js
@@ -126,6 +126,7 @@ describe('Complete login flow', function() {
     return bootstrap({})
     .then(function(app) {
       return app.loginRedirect({
+        pkce: false,
         nonce: NONCE
       });
     })