+++ test/support/oauthUtil.js
@@ -167,6 +167,7 @@ oauthUtil.setup = function(opts) {
     authClient = opts.authClient;
   } else {
     authClient = new OktaAuth({
+      pkce: false,
       issuer: 'https://auth-js-test.okta.com'
     });
   }
@@ -357,6 +358,7 @@ oauthUtil.setupPopup = function(opts) {
 
 oauthUtil.setupRedirect = function(opts) {
   var client = new OktaAuth(Object.assign({
+    pkce: false,
     issuer: 'https://auth-js-test.okta.com',
     clientId: 'NPSfOkH5eZrTy8PMDlvx',
     redirectUri: 'https://example.com/redirect'
@@ -385,6 +387,7 @@ oauthUtil.setupRedirect = function(opts) {
 
 oauthUtil.setupParseUrl = function(opts) {
   var client = new OktaAuth({
+    pkce: false,
     issuer: 'https://auth-js-test.okta.com',
     clientId: 'NPSfOkH5eZrTy8PMDlvx',
     redirectUri: 'https://example.com/redirect'
@@ -456,6 +459,7 @@ oauthUtil.setupParseUrl = function(opts) {
 oauthUtil.setupSimultaneousPostMessage = function() {
   // Create client
   var client =  new OktaAuth({
+    pkce: false,
     issuer: 'https://auth-js-test.okta.com',
     clientId: 'NPSfOkH5eZrTy8PMDlvx',
     redirectUri: 'https://example.com/redirect'